Complete task 1

+++ src/DataTypes.java
@@ -1,11 +1,11 @@
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(int[] numbers) {
-        int s = 0; // variable to accumulate the sum in!
+        long s = 0; // variable to accumulate the sum in!
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
         }
         return s;
     }
+
 }
